+++ src/RestApi/Emulator/IEmulator.ts
@@ -1,0 +1,10 @@
+export interface IEmulatorCreatable {
+	organizationUid: string;
+}
+
+export default interface IEmulator {
+	uid: string;
+	duid: string;
+	name: string;
+	createdAt: Date;
+}