+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added method to get list of emulators for given account and method to create new emulator
+
 ## [1.4.2] - 2020-10-30
 ### Deprecated
 - `Api` constructor options was renamed. `accountAuth.accountId` => `accountAuth.tokenId` & `accountAuth.secret` => `accountAuth.token`.